+++ tests/toisto/model/test_filter.py
@@ -20,7 +20,7 @@ class FilterTest(ToistoTestCase):
         self.argument_parser = ArgumentParser()
         self.foo = self.create_concept("foo", {})
         self.bar = self.create_concept("bar", {})
-        self.bar_id = ConceptId("bar")
+        self.bar_id = self.bar.concept_id
         self.baz_id = ConceptId("baz")
         self.concepts = {self.foo, self.bar}
 
@@ -56,10 +56,10 @@ class FilterTest(ToistoTestCase):
         self.assertRaises(SystemExit, self.filter_concepts, selected_concepts=[self.baz_id])
         self.assertIn("Concept 'baz' not found", sys_stderr_write.call_args_list[1][0][0])
 
-    def test_add_hypernyms_and_hyponyms_of_selected_concepts(self):
-        """Test that the hypernyms and hyponyms of selected concepts are added."""
-        little_bar = self.create_concept("little bar", dict(hypernym="bar"))
-        little_beer_bar = self.create_concept("little beer bar", dict(hypernym="little bar"))
+    def test_add_hyponyms_of_selected_concepts_recursively(self):
+        """Test that the hyponyms of selected concepts are added, recursively."""
+        little_bar = self.create_concept("little bar", dict(hypernym=self.bar_id))
+        little_beer_bar = self.create_concept("little beer bar", dict(hypernym=little_bar.concept_id))
         self.assertEqual(
             {self.bar, little_bar, little_beer_bar},
             self.filter_concepts(
@@ -67,19 +67,41 @@ class FilterTest(ToistoTestCase):
             ),
         )
 
-    def test_add_holonyms_and_meronyms_of_selected_concepts(self):
-        """Test that the holonyms and meronyms of selected concepts are added."""
-        bar_part = self.create_concept("bar part", dict(holonym="bar"))
-        bar_part_part = self.create_concept("bar part part", dict(holonym="bar part"))
+    def test_do_not_add_hypernyms_of_selected_concepts(self):
+        """Test that the hypernyms of selected concepts are not added."""
+        little_bar = self.create_concept("little bar", dict(hypernym=self.bar_id))
+        little_beer_bar = self.create_concept("little beer bar", dict(hypernym=little_bar.concept_id))
+        self.assertEqual(
+            {little_beer_bar},
+            self.filter_concepts(
+                concepts=self.concepts | {little_bar, little_beer_bar}, selected_concepts=[little_beer_bar.concept_id]
+            ),
+        )
+
+    def test_add_meronyms_of_selected_concepts_recursively(self):
+        """Test that the holonyms and meronyms of selected concepts are added, recursively."""
+        bar_part = self.create_concept("bar part", dict(holonym=self.bar_id))
+        bar_part_part = self.create_concept("bar part part", dict(holonym=bar_part.concept_id))
         self.assertEqual(
             {self.bar, bar_part, bar_part_part},
             self.filter_concepts(concepts=self.concepts | {bar_part, bar_part_part}, selected_concepts=[self.bar_id]),
         )
 
+    def test_do_not_add_holonyms_of_selected_concepts(self):
+        """Test that the holonyms of selected concepts are not added."""
+        bar_part = self.create_concept("bar part", dict(holonym=self.bar_id))
+        bar_part_part = self.create_concept("bar part part", dict(holonym=bar_part.concept_id))
+        self.assertEqual(
+            {bar_part_part},
+            self.filter_concepts(
+                concepts=self.concepts | {bar_part, bar_part_part}, selected_concepts=[bar_part_part.concept_id]
+            ),
+        )
+
     def test_add_hyponyms_of_meronyms_of_selected_concepts(self):
         """Test that hyponyms of meronyms of selected concepts are added."""
-        bar_part = self.create_concept("bar part", dict(holonym="bar"))
-        little_bar_part = self.create_concept("little bar part", dict(hypernym="bar part"))
+        bar_part = self.create_concept("bar part", dict(holonym=self.bar_id))
+        little_bar_part = self.create_concept("little bar part", dict(hypernym=bar_part.concept_id))
         self.assertEqual(
             {self.bar, bar_part, little_bar_part},
             self.filter_concepts(concepts=self.concepts | {bar_part, little_bar_part}, selected_concepts=[self.bar_id]),
@@ -87,8 +109,8 @@ class FilterTest(ToistoTestCase):
 
     def test_add_meronyms_of_hyponyms_of_selected_concepts(self):
         """Test that meronyms of hyponyms of selected concepts are added."""
-        little_bar = self.create_concept("little bar", dict(hypernym="bar"))
-        little_bar_part = self.create_concept("little bar part", dict(holonym="little bar"))
+        little_bar = self.create_concept("little bar", dict(hypernym=self.bar_id))
+        little_bar_part = self.create_concept("little bar part", dict(holonym=little_bar.concept_id))
         self.assertEqual(
             {self.bar, little_bar, little_bar_part},
             self.filter_concepts(
@@ -98,28 +120,25 @@ class FilterTest(ToistoTestCase):
 
     def test_add_antonyms_of_selected_concepts(self):
         """Test that antonyms of selected concepts are added."""
-        anti_foo = self.create_concept("anti foo", dict(antonym="foo"))
+        anti_foo = self.create_concept("anti foo", dict(antonym=self.foo.concept_id))
         self.assertEqual(
             {self.foo, anti_foo},
-            self.filter_concepts(concepts=self.concepts | {anti_foo}, selected_concepts=[ConceptId("anti foo")]),
+            self.filter_concepts(concepts=self.concepts | {anti_foo}, selected_concepts=[anti_foo.concept_id]),
         )
 
     def test_add_concepts_that_have_selected_concepts_as_root(self):
         """Test that the concepts that have a selected concept as root are added."""
-        little_bar = self.create_concept("little bar", dict(roots="bar"))
+        little_bar = self.create_concept("little bar", dict(roots=self.bar_id))
         self.assertEqual(
             {self.bar, little_bar},
             self.filter_concepts(concepts=self.concepts | {little_bar}, selected_concepts=[self.bar_id]),
         )
 
-    def test_add_concepts_involved_with_and_by_selected_concepts(self):
-        """Test that concepts involved with and by selected concepts are added."""
-        zoo = self.create_concept("zoo", dict(involves="bar"))
+    def test_add_concepts_involved_by_selected_concepts_recursively(self):
+        """Test that concepts involved by selected concepts are added, recursively."""
+        zoo = self.create_concept("zoo", dict(involves=self.bar_id))
+        baz = self.create_concept("baz", dict(involves=zoo.concept_id))
         self.assertEqual(
-            {self.bar, zoo},
-            self.filter_concepts(concepts=self.concepts | {zoo}, selected_concepts=[self.bar_id]),
-        )
-        self.assertEqual(
-            {self.bar, zoo},
-            self.filter_concepts(concepts=self.concepts | {zoo}, selected_concepts=[ConceptId("zoo")]),
+            {self.bar, baz, zoo},
+            self.filter_concepts(concepts=self.concepts | {baz, zoo}, selected_concepts=[baz.concept_id]),
         )