+++ CHANGELOG.md
@@ -11,11 +11,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Accept an initial capital letter when both the question and the answer of a quiz are in lower case. Fixes [#671](https://github.com/fniessink/toisto/issues/671).
 - Use single quotes when referring to questions and answers in user feedback. When feedback ends with a quoted question or answer, only add a period if the quoted question or answer does not already end with punctuation. Fixes [#675](https://github.com/fniessink/toisto/issues/675).
 - Allow for omitting the article when the source language is Dutch and the answer has a capital. For example, when the correct answer was "het Engels", answering "Engels" would be marked as incorrect. Fixes [#680](https://github.com/fniessink/toisto/issues/680).
+- When concepts have roots that also have roots, not only quiz the direct root concepts before the compound concept but also the roots of the roots. For example, one of the roots of "last weekend" would be "weekend", which in turn has "week" and "end" as roots. Fixes [#683](https://github.com/fniessink/toisto/issues/683).
 
 ### Added
 
 - Add support for the present perfect tense. Closes [#632](https://github.com/fniessink/toisto/issues/632).
 - When showing examples, also show the meaning of the examples. Closes [#638](https://github.com/fniessink/toisto/issues/638).
+- Clarify in the documentation which concept relations are recursive. Closes [#683](https://github.com/fniessink/toisto/issues/683).
 - Add several concepts.
 
 ## 0.18.1 - 2024-04-09